+++ source/src/calceff.cpp
@@ -6,21 +6,23 @@
 // ============================================================================
 #include <cmath>
 #include <limits>
-#include <tuple>
+#include <array>
 // ============================================================================
 // GSL 
 // ============================================================================
-#include "gsl/gsl_roots.h"
-#include "gsl/gsl_multiroots.h"
+#include "gsl/gsl_min.h"
 // ============================================================================
 // Ostap
 // ============================================================================
 #include "Ostap/Binomial.h"
 #include "Ostap/MoreMath.h"
+#include "Ostap/LinAlg.h"
 // ============================================================================
 // Local 
 // ============================================================================
 #include "local_math.h"
+#include "local_gsl.h"
+#include "GSL_sentry.h"
 // ============================================================================
 /** @file
  *  implementation of functions from file Ostap/Binomial.h
@@ -30,77 +32,25 @@
 // ============================================================================
 namespace
 {
-  // // ==========================================================================
-  // typedef std::tuple<unsigned long,unsigned long,double> PARAMS ;
-  // // ==========================================================================
-  // double  B1Df ( double x , void* p )
-  // {
-  //   const PARAMS* params = static_cast<PARAMS*> ( p ) ;
-  //   const unsigned long A = std::get<0> ( params ) ;
-  //   const unsigned long B = std::get<1> ( params ) ;
-  //   const double        P = std::get<2> ( params ) ;
-  //   return Ostap::Math::beta_inc ( 1.0 * A , 1.0 * B , x ) - P ;
-  // }
-  // double B1Ddf ( double x , void* p )
-  // {
-  //   const PARAMS* params = static_cast<PARAMS*> ( p ) ;
-  //   const unsigned long A = std::get<0> ( params ) ;
-  //   const unsigned long B = std::get<1> ( params ) ;
-  //   return Ostap::Math::dbeta_inc ( 1.0 * A , 1.0 * B , x ) ;
-  // }
-  // void   B1Dfdf
-  // ( double  x  ,
-  //   void*   p  , 
-  //   double* f  ,
-  //   double* df )
-  // {
-  //   const PARAMS* params = static_cast<PARAMS*> ( p ) ;
-  //   const unsigned long A = std::get<0> ( params ) ;
-  //   const unsigned long B = std::get<1> ( params ) ;
-  //   const double        P = std::get<2> ( params ) ;
-  //   *f  = Ostap::Math::beta_inc  ( 1.0 * A , 1.0 * B , x ) - P ;
-  //   *df = Ostap::Math::dbeta_inc ( 1.0 * A , 1.0 * B , x )     ;
-  // }
-  // // ================================================================================
-  
-  // int bayes    ( gsl_vector* x , void* p , gsl_vector* f )
-  // {
-  //   const double alpha  = gsl_vector_get ( x , 0 ) ; // parameter alpha 
-  //   const double beta   = gsl_vector_get ( x , 1 ) ; // parameter   beta 
-  //   const double lambda = gsl_vector_get ( x , 2 ) ; // Lagrange  muliplier 
-  //   //
-  //   std::array<double,3> params = static_cast<std::array<double,3> > ( p ) ;
-  //   //
-  //   const double a = params [ 0 ] ;
-  //   const double b = params [ 1 ] ;
-  //   const double P = params [ 2 ] ;
-  //   //
-  //   gsl_vector_set ( f , 0 , 1 + lambda * Ostap::Math::dbeta_inc ( a , b , alpha ) ) ;
-  //   gsl_vector_set ( f , 1 , 1 + lambda * Ostap::Math::dbeta_inc ( a , b , beta  ) ) ;
-  //   gsl_vector_set ( f , 2 ,
-  //                    Ostap::Math::beta_inc ( a , b , beta  ) -
-  //                    Ostap::Math::beta_inc ( a , b , alpha ) - P ) ;
-  //   //
-  //   return GSL_SUCCESS ;
-  // }
-  // // ==========================================================================
-  // int bayes_fd ( gsl_vector* x , void* p , gsl_matrix* J )
-  // {
-    
-  //   const double alpha  = gsl_vector_get ( x , 0 ) ; // parameter alpha 
-  //   const double beta   = gsl_vector_get ( x , 1 ) ; // parameter   beta 
-  //   const double lambda = gsl_vector_get ( x , 2 ) ; // Lagrange  muliplier 
-  //   //
-  //   std::array<double,3> params = static_cast<std::array<double,3> > ( p ) ;
-  //   //
-  //   const double a = params [ 0 ] ;
-  //   const double b = params [ 1 ] ;
-  //   const double P = params [ 2 ] ;
-  //   //
-
-  // }
-    
-                 
+  // ==========================================================================
+  typedef std::array<double,3> PARAMS ;
+  // ==========================================================================
+  //
+  double beta_minus_alpha
+  ( const double x , void* p )
+  {
+    //
+    const PARAMS* params = static_cast<PARAMS*> ( p ) ;
+    //
+    const double alpha = ( *params ) [ 0 ] ;
+    const double beta  = ( *params ) [ 1 ] ;
+    const double P     = ( *params ) [ 2 ] ;
+    // 
+    const double CDF = Ostap::Math::beta_cdf      ( x       , alpha , beta ) ;
+    const double Q   = Ostap::Math::beta_quantile ( CDF + P , alpha , beta ) ;
+    //
+    return Q - x ;
+  } ;    
   // ==========================================================================  
 }
 // ============================================================================
@@ -117,76 +67,92 @@ Ostap::Math::bayes_interval
   const unsigned long rejected  ,
   const double        conflevel )
 {
-  if ( 0 == accepted && 0 == rejected ) { return std::make_pair ( 0.0 , 1.0 ) ; }
+  if ( 0 == accepted && 0 == rejected              ) { return std::make_pair ( 0.0 , 1.0 ) ; }
+  if ( 1 <= conflevel || s_equal ( conflevel , 1 ) ) { return std::make_pair ( 0.0 , 1.0 ) ; }
   //
-  if ( 1 <= conflevel ) { return std::make_pair ( 0.0 , 1.0 ) ; }
   const double a = accepted ;
-  const double p = a / ( a + rejected ) ;
-  if ( 0 >= conflevel  ) { return std::make_pair ( p , p )    ; }
+  const double e = a / ( a + rejected ) ;
+  if ( 0 >= conflevel || s_zero ( conflevel ) ) { return std::make_pair ( e , e ) ; }
+  //
+  if ( !accepted ) { return std::make_pair ( 0.0 , 1 - std::pow ( 1 - conflevel , 1.0 / ( rejected + 1 ) )       ) ; }
+  if ( !rejected ) { return std::make_pair (           std::pow ( 1 - conflevel , 1.0 / ( accepted + 1 ) ) , 1.0 ) ; }  
+  ///
+  const double par_alpha = accepted + 1 ;
+  const double par_beta  = rejected + 1 ;
+  //
+  // The interval is:
+  //   0<= alpha <= e <= beta <= 1, such that  CDF ( beta ) - CDF ( alpha ) == cnpflevel 
+  //
+  /// CDF for Beta-distribution
+  auto CDF = [par_alpha,par_beta] ( const double x ) -> double
+  { return Ostap::Math::beta_cdf      ( x , par_alpha , par_beta ) ; } ;
+  /// Quantile for Beta-distribution  
+  auto Q   = [par_alpha,par_beta] ( const double p ) -> double
+  { return Ostap::Math::beta_quantile ( p , par_alpha , par_beta   ) ; } ;       
+  ///
+
+
+  // use GSL: 
+  // Ostap::Math::GSL::GSL_Error_Handler sentry ;
+  Ostap::Utils::GslIgnore  sentry { true } ;
+
+  //
+  // The [alpha,beta] interval is:
+  //   0 <= alpha <= e <= beta <= 1,
+  //   such that  CDF ( beta ) - CDF ( alpha ) == cnpflevel
+  // due to this constrain for each (fixed) alpha beta is a functiopn of alpha
+  // and we need to find such alphs taht minimizes the length of [alpha,beta] interval     
   //
   
-  if ( !accepted ) { return std::make_pair ( 0.0 , 1 - std::pow ( 1 - conflevel , 1.0 / ( rejected + 1 ) )       )   ; }
-  
-  if ( !rejected ) { return std::make_pair (           std::pow ( 1 - conflevel , 1.0 / ( accepted + 1 ) ) , 1.0 ) ; }  
-  
-  // if ( !accepted )
-  //   {        
-  //     //
-  //     PARAMS params { accepted , rejected , !accepted ? ( 1 - conflevel ) : conflevel }  ;
-  //     // brent 
-  //     gsl_function  F ;
-  //     F  .f  = &B1D     ;
-      
-  //     // Brent 
-  //     const gsl_root_fsolver_type* TB ;
-  //     gsl_root_fsolver*            sb ;      
-  //     TB = gsl_root_fsolver_brent (    ) ;
-  //     sb = gsl_root_fsolver_alloc ( TB ) ;
-  //     gsl_root_fsolver_set ( sb , &F , 0 , 1 ) ; 
-  //     //
-
-  //     // ======================================================================
-  //     // several iterations of Brent
-  //     // ======================================================================
-  //     int biter   = 0 ;
-  //     int bstatus = 0 ;
-  //     double xlow = 0 ;
-  //     double xigh = 1 ;
-      
-  //     for ( int biter = 0 ; biter < 5 ; ++biter )
-  //       {
-  //         int bstatus = gsl_root_fsolver_iterate ( sb ) ;
-      
-  //       }
-      
-  //     // switch to Newton 
-      
-  //     // Newton 
-  //     const gsl_root_fdfsolver_type* T2 ;
-  //     gsl_root_fdfsolver*            s2 ;      
-  //     T2 = gsl_root_fdfsolver_newton (    ) ;
-  //     s2 = gsl_root_fdfsolver_alloc  ( T2 ) ;
-  //     //
-  //     gsl_root_fdfsolver_set ( s , &F , 0 , 1 ) ; 
-
-
-      
-  //     // newtom 
-  //     gsl_function_fdf FDF ;
-  //     FDF.f   = &B1Df   ;
-  //     FDF.df  = &B1Ddf  ;
-  //     FDF.fdf = &B1Dfdf ;
-  //     //
-  //     F  .params = &params ;
-  //     FDF.params = &params ;
-  //     //
-  //     //
-      
-  //   }
-  
-  return std::make_pair ( p , p  ) ;  
+  const double e_CDF     = CDF ( e ) ;
+  const double alpha_min = e_CDF <= conflevel ? 0.0 : Q ( e_CDF - conflevel ) ;
+  const double alpha_max = std::min ( e , Q ( 1 - conflevel  ) ) ; 
+  ///
+  // Jump to GSL-world 
+  //
+  PARAMS params { par_alpha  , par_beta , conflevel } ;
+  // ==========================================================================  
+  // 
+  const gsl_min_fminimizer_type    *T ;
+  gsl_min_fminimizer               *s ;
+  //
+  gsl_function F ;
+  F.function   = &beta_minus_alpha    ; 
+  F.params     = &params              ;
+  //
+  T = gsl_min_fminimizer_brent        ;
+  s = gsl_min_fminimizer_alloc ( T )  ;
+  //
+  gsl_min_fminimizer_set ( s, &F,  0.5 * ( alpha_min + alpha_max ) , alpha_min , alpha_max );
+  //
+  const double min_e     = 0.001 * std::min ( e         , 1 - e         ) ;
+  const double min_c     = 0.001 * std::min ( conflevel , 1 - conflevel ) ;
+  static const double DX = std::clamp ( std::min ( min_e , min_c ) , 1.e-8 , 1.e-7 ) ; 
+  //
+  double low   = alpha_min ;
+  double high  = alpha_max ;
+  double alpha = 0.5 * ( alpha_min + alpha_max ) ;
+  for ( unsigned short i = 0 ; i < 50 ; ++i )
+    {
+      int status = gsl_min_fminimizer_iterate   ( s ) ;
+      //
+      if ( status ) { gsl_error ( "Error from gsl_min_fminimizer_iterate" , __FILE__ , __LINE__ , status ) ; }
+      //
+      alpha      = gsl_min_fminimizer_x_minimum ( s ) ;
+      low        = gsl_min_fminimizer_x_lower   ( s ) ;
+      high       = gsl_min_fminimizer_x_upper   ( s ) ;
+      //
+      if ( GSL_SUCCESS == gsl_min_test_interval ( low , high , DX , DX ) ) { break ; } 
+      else if ( std::abs ( high - low ) <= DX                            ) { break ; }          
+      //
+    }
+  //
+  gsl_min_fminimizer_free( s ) ;
+  // upper edge of the interval 
+  const double beta = Q ( CDF ( alpha ) + conflevel ) ;
+  // result 
+  return std::make_pair ( alpha , beta ) ;  
 }
-
 // ============================================================================
 //                                                                      The END 
 // ============================================================================